Pull the next 5 tweets and relevant metadata
SERVER

create a new TwitterApi client using a bearer token

set client to read only

create a new user object with usersByUsernames, then get the id of that user & set it to myTwitterId

With this new Id property, we are able to GET the last 5 tweets from the user with the userTimeline function built into v2.

Setting the max results to 5, we limit it to 5.