Processing requests from client side to server side
----- SearchBar.jsx -----

First, we want to check if the incoming data is non-empty

If it is not empty, then we do a POST request to the search route with our POST payload as inputText. Essentially, this inputText is the username that the user submits into the search bar

----- server.js -----

line 28
Server receives this post request from the /search route. Doubling on error catching; if the payload (parcel) does not exist, we send back a 400.

If it exists, we send back an ok response (200), and call the fetchingTweets function with parcel as an arg.

Remember, parcel is the usernameQuery on the client-side. It's what the user has submitted from the search bar.

I refactored the fetchingTweets into a separate function since it would be reusable - the only argument it takes is the usernameQuery/parcel argument.

Once the twitter-api client fetches the data and hydrates the JSON object to be returned (line 44 - line 173), we call getUser, which passes the tweets object and the usernameQuery.

getUser creates a GET route for the usernameQuery AT route /tweets/usernameQuery and puts the data there.

e.g. the user types "POTUS" into the search bar, submits. After twitter-api hydrates, getUser creates a new route /tweets/POTUS with the hydrated data.

----- BACK TO SearchBar.jsx -----

Once the route is created and the requested data is posted on that route, we must fetch for that data.

This means calling fetchTweetData with an argument of inputText, which is assumed to be the user's query.

fetchTweetData() is located in App.jsx

----- App.jsx -----

Once fetchTweetData(inputText) is called, the browser will fetch the data from the dynamically generated route, http://localhost:5000/tweets/POTUS