Error handling - read desc
Input that the user enters needs to be validated before it can be fetched

isValidExpression checks if user's inputText is valid. This expression bound is defined by Twitter when accessing their API

if the input text is truthy but the expression is not valid, we must display a relevant error message - that usernames must be within expression bounds.